LayoutTests/http/tests/workers/service/registration-task-queue-scheduling-1-expected.txt

CONSOLE MESSAGE: line 48: Original window resolved successfully (unexpected)
